+++ packages/app-angular/src/app/directives/customValueAccessors.directives.ts
@@ -11,14 +11,15 @@ import { BaseCustomValueAccessor } from './baseCustomValueAccessor.directives';
 			multi: true,
 		},
 	],
+	standalone: false,
 })
 export class OntarioInputValueAccessor extends BaseCustomValueAccessor {
 	@HostListener('inputOnChange', ['$event.detail.value'])
-	_handleChange(value: string) {
-		this.onChange(value);
+	_handleChange(value?: string) {
+		this.onChange(value ?? '');
 	}
 
-	@HostListener('onBlur', ['$event.target'])
+	@HostListener('onBlur')
 	_handleBlurEvent() {
 		this.onTouched();
 	}
@@ -33,14 +34,15 @@ export class OntarioInputValueAccessor extends BaseCustomValueAccessor {
 			multi: true,
 		},
 	],
+	standalone: false,
 })
 export class OntarioTextareaValueAccessor extends BaseCustomValueAccessor {
 	@HostListener('inputOnChange', ['$event.detail.value'])
-	_handleChange(value: string) {
-		this.onChange(value);
+	_handleChange(value?: string) {
+		this.onChange(value ?? '');
 	}
 
-	@HostListener('onBlur', ['$event.target'])
+	@HostListener('onBlur')
 	_handleBlurEvent() {
 		this.onTouched();
 	}
@@ -55,6 +57,7 @@ export class OntarioTextareaValueAccessor extends BaseCustomValueAccessor {
 			multi: true,
 		},
 	],
+	standalone: false,
 })
 export class OntarioCheckboxesValueAccessor extends BaseCustomValueAccessor {
 	@HostListener('checkboxOnChange', ['$event.detail'])
@@ -62,7 +65,7 @@ export class OntarioCheckboxesValueAccessor extends BaseCustomValueAccessor {
 		this.onChange(event);
 	}
 
-	@HostListener('checkboxOnBlur', ['$event'])
+	@HostListener('checkboxOnBlur')
 	_handleBlurEvent() {
 		this.onTouched();
 	}
@@ -77,6 +80,7 @@ export class OntarioCheckboxesValueAccessor extends BaseCustomValueAccessor {
 			multi: true,
 		},
 	],
+	standalone: false,
 })
 export class OntarioRadioButtonsValueAccessor extends BaseCustomValueAccessor {
 	@HostListener('radioOnChange', ['$event.detail'])
@@ -84,7 +88,7 @@ export class OntarioRadioButtonsValueAccessor extends BaseCustomValueAccessor {
 		this.onChange(event.detail);
 	}
 
-	@HostListener('radioOnBlur', ['$event'])
+	@HostListener('radioOnBlur')
 	_handleBlurEvent() {
 		this.onTouched();
 	}